pyproject.toml
[tool.poetry]
name = "PhotoFriends"
version = "0.0.1"
description = "Share photo inside a group of friends"
authors = ["Benjamin <benji81@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
gunicorn = "^20.1.0"
Pillow = "^8.4.0"
django-bootstrap5 = "^2.1.2"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black = "^21.7b0"
flake8 = "^3.9.2"
coverage = "^5.5"
mypy = "^0.910"
isort = "^5.9.3"
pre-commit = "^2.14.0"
pytest = "^6.2.4"
requests = "^2.26.0"
django-stubs = "^1.8.0"
ipython = "^7.26.0"
pytest-django = "^4.4.0"
pytest-cov = "^2.12.1"
pylint = "^2.11.1"
pylint-django = "^2.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.tox
    | \.venv
    | _build
    | build
    | dist
    | docs
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
combine_as_imports = true
lines_after_imports = 2
force_sort_within_sections = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
DJANGO_SETTINGS_MODULE = "photofriends.settings"
python_files = "tests.py test_*.py *_tests.py"
